+++ tardis/io/model/parse_packet_source_configuration.py
@@ -1,8 +1,8 @@
 from astropy import units as u
 
-from tardis.transport.montecarlo.packet_source import (
+from tardis.transport.montecarlo.packet_source.black_body_relativistic import BlackBodySimpleSourceRelativistic
+from tardis.transport.montecarlo.packet_source.black_body import (
     BlackBodySimpleSource,
-    BlackBodySimpleSourceRelativistic,
 )
